Set of Complements

ComplementSet class for symbolic complement of mathematical sets.

Details

The purpose of this class is to provide a symbolic representation for the complement of sets that cannot be represented in a simpler class. Whilst this is not an abstract class, it is not recommended to construct this class directly but via the set operation methods.

See also

Super classes

set6::Set -> set6::SetWrapper -> ComplementSet

Active bindings

elements

Returns the elements in the object.

length

Returns the number of elements in the object.

addedSet

For the ComplementSet wrapper, X-Y, returns the set X.

subtractedSet

For the ComplementSet wrapper, X-Y, returns the set Y.

Methods

Public methods

ComplementSet$new()

ComplementSet$strprint()

ComplementSet$contains()

ComplementSet$clone()

Inherited methods

set6::Set$add()

set6::Set$print()

set6::Set$remove()

set6::Set$summary()

set6::SetWrapper$equals()

set6::SetWrapper$isSubset()

Method new()

Create a new ComplementSet object. It is not recommended to construct this class directly.

Usage

ComplementSet$new(addset, subtractset, lower = NULL, upper = NULL, type = NULL)

Arguments

addset

Set to be subtracted from.

subtractset

Set to subtract.

lower

lower bound of new object.

upper

upper bound of new object.

type

closure type of new object.

Returns

A new ComplementSet object.

Method strprint()

Creates a printable representation of the object.

Usage

ComplementSet$strprint(n = 2)

Arguments

n

numeric. Number of elements to display on either side of ellipsis when printing.

Returns

A character string representing the object.

Method contains()

Tests if elements x are contained in self.

Usage

ComplementSet$contains(x, all = FALSE, bound = FALSE)

Arguments

x

Set or vector of Sets.

all

logical. If FALSE tests each x separately. Otherwise returns TRUE only if all x pass test.

bound

logical

Returns

If all == TRUE then returns TRUE if all x are contained in self, otherwise FALSE. If all == FALSE returns a vector of logicals corresponding to the length of x, representing if each is contained in self. If bound == TRUE then an element is contained in self if it is on or within the (possibly-open) bounds of self, otherwise TRUE only if the element is within self or the bounds are closed.

Method clone()

The objects of this class are cloneable with this method.

Usage

ComplementSet$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.