Set of Extended Real Numbers

The mathematical set of extended real numbers, defined as the union of the set of reals with \(\pm\infty\). i.e. $$R \cup \\{-\infty, \infty\\}$$ where \(R\) is the set of reals.

See also

Super classes

set6::Set -> set6::Interval -> set6::SpecialSet -> set6::Reals -> ExtendedReals

Methods

Public methods

ExtendedReals$new()

ExtendedReals$clone()

Inherited methods

set6::Set$add()

set6::Set$print()

set6::Set$remove()

set6::Set$summary()

set6::Interval$contains()

set6::Interval$equals()

set6::Interval$isSubinterval()

set6::Interval$isSubset()

set6::SpecialSet$strprint()

Method new()

Create a new ExtendedReals object.

Usage

ExtendedReals$new()

Returns

A new ExtendedReals object.

Method clone()

The objects of this class are cloneable with this method.

Usage

ExtendedReals$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.