Set of Rational Numbers

The mathematical set of rational numbers, defined as the set of numbers that can be written as a fraction of two integers. i.e. $$\\{\frac{p}{q} \ : \ p,q \ \in \ Z, \ q \ne 0 \\}$$ where \(Z\) is the set of integers.

Details

The $contains method does not work for the set of Rationals as it is notoriously difficult/impossible to find an algorithm for determining if any given number is rational or not. Furthermore, computers must truncate all irrational numbers to rational numbers.

See also

Super classes

set6::Set -> set6::Interval -> set6::SpecialSet -> Rationals

Methods

Public methods

Rationals$new()

Rationals$clone()

Inherited methods

set6::Set$add()

set6::Set$print()

set6::Set$remove()

set6::Set$summary()

set6::Interval$contains()

set6::Interval$equals()

set6::Interval$isSubinterval()

set6::Interval$isSubset()

set6::SpecialSet$strprint()

Method new()

Create a new Rationals object.

Usage

Rationals$new(lower = -Inf, upper = Inf, type = "()")

Arguments

lower

numeric. Where to start the set. Advised to ignore, used by child-classes.

upper

numeric. Where to end the set. Advised to ignore, used by child-classes.

type

character Set closure type. Advised to ignore, used by child-classes.

Returns

A new Rationals object.

Method clone()

The objects of this class are cloneable with this method.

Usage

Rationals$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.