Abstract Class for Special Sets

The 'special sets' are the group of sets that are commonly used in mathematics and are thus given their own names.

Details

This is an abstract class and should not be constructed directly. Use listSpecialSets to see the list of implemented special sets.

Super classes

set6::Set -> set6::Interval -> SpecialSet

Methods

Public methods

SpecialSet$new()

SpecialSet$strprint()

SpecialSet$clone()

Inherited methods

set6::Set$add()

set6::Set$print()

set6::Set$remove()

set6::Set$summary()

set6::Interval$contains()

set6::Interval$equals()

set6::Interval$isSubinterval()

set6::Interval$isSubset()

Method new()

SpecialSet is an abstract class, the constructor cannot be used directly.

Usage

SpecialSet$new(lower = -Inf, upper = Inf, type = "()", class = "numeric")

Arguments

lower

defines the lower bound of the interval.

upper

defines the upper bound of the interval.

type

defines the interval closure type.

class

defines the interval class.

Method strprint()

Creates a printable representation of the object.

Usage

SpecialSet$strprint(n = NULL)

Arguments

n

ignored, added for consistency.

Returns

A character string representing the object.

Method clone()

The objects of this class are cloneable with this method.

Usage

SpecialSet$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.