Set Properties Class

Used to store the properties of a Set. Though this is not an abstract class, it should never be constructed outside of the Set constructor.

Active bindings

closure

Returns the closure of the Set. One of "open", "half-open", or "closed."

countability

Returns the countability of the Set. One of "countably finite", "countably infinite", or "uncountable".

cardinality

Returns the cardinality of the Set. Either an integer if the Set is countably finite, Aleph0 if countably infinite, or a Beth number.

empty

Returns if the Set is empty or not. TRUE if the Set cardinality is 0, FALSE otherwise.

singleton

Returns if the Set is a singleton or not. TRUE if the Set cardinality is 1, FALSE otherwise.

Methods

Public methods

Properties$new()

Properties$print()

Properties$strprint()

Properties$clone()

Method new()

Creates a new Properties object.

Usage

Properties$new(closure = character(0), cardinality = NULL)

Arguments

closure

One of "open", "half-open", or "closed."

cardinality

If non-NULL then either an integer, "Aleph0", or a Beth number.

Returns

A new Properties object.

Method print()

Prints the Properties list.

Usage

Properties$print()

Returns

Prints Properties list to console.

Method strprint()

Creates a printable representation of the Properties.

Usage

Properties$strprint()

Returns

A list of properties.

Method clone()

The objects of this class are cloneable with this method.

Usage

Properties$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.