assert/check/test/Interval

Validation checks to test if a given object is an R6 Interval.

testInterval(object, errormsg = "This is not an R6 Interval object")

checkInterval(object, errormsg = "This is not an R6 Interval object")

assertInterval(object, errormsg = "This is not an R6 Interval object")

Arguments

object
object to test

errormsg
error message to overwrite default if check fails

Value

If check passes then assert returns object invisibly and test/check return TRUE. If check fails, assert stops code with error, check returns an error message as string, and test returns FALSE.

Examples

testInterval(Set$new(2, 3))

#> [1] FALSE

testInterval(list(Set$new(2), Set$new(3)))

#> [1] FALSE

testInterval(Tuple$new(2, 3))

#> [1] FALSE

testInterval(Interval$new())

#> [1] TRUE

testInterval(FuzzySet$new(2, 0.1))

#> [1] FALSE

testInterval(FuzzyTuple$new(2, 0.1))

#> [1] FALSE

testInterval(ConditionalSet$new(function(x) x == 0))

#> [1] FALSE